pyproject.toml
[tool.poetry]
name = "authorization by phone number"
version = "0.1.0"
description = ""
authors = ["ika <mamina747@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
black = "^24.10.0"
djangorestframework-simplejwt = "^5.3.1"
smsaero-api = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
